services/fleet-service/pyproject.toml
[project]
name = "fleet-service"
version = "0.1.0"
description = "Fleet Service — Canonical vehicle and trailer master data, lifecycle, technical specs, and asset validation (v1.5 spec)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-ulid>=3.0.0",
    "httpx>=0.27.0",
    "prometheus-client>=0.20.0",
    "PyJWT>=2.10.1",
    "confluent-kafka>=2.13.2",
]

[project.scripts]
fleet-api = "fleet_service.entrypoints.api:main"
fleet-worker = "fleet_service.entrypoints.worker:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "testcontainers[postgres]>=4.0.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.ruff.lint.isort]
known-first-party = ["fleet_service"]

[tool.mypy]
python_version = "3.12"
strict = true
